Move the workbench reducers out of state/mod.rs to hold the size gate (#626)
Merging main pushed src/state/mod.rs to 1007 lines, over the 1000 hard limit:
main grew it and the workbench reducers were the last straw. The reducers are
self-contained and already have a home beside the rest of the workbench state,
so they move there rather than the limit moving.
